@@ -1,12 +1,11 @@
-﻿namespace CodeAnalyzer.Tests.Analysis.Cohesion.TestClasses
-{
-    public class ClassWithNoMethod
-    {
-        private readonly string _field;
+﻿namespace CodeAnalyzer.Tests.Analysis.Cohesion.TestClasses;
 
-        public ClassWithNoMethod(string field)
-        {
-            _field = field;
-        }
+public class ClassWithNoMethod
+{
+    private readonly string _field;
+
+    public ClassWithNoMethod(string field)
+    {
+        _field = field;
     }
 }
