@@ -1,14 +1,13 @@
-﻿namespace CodeAnalyzer.Tests.Analysis.Cohesion.TestClasses
+﻿namespace CodeAnalyzer.Tests.Analysis.Cohesion.TestClasses;
+
+public class ClassWithNoPublicMethod
 {
-    public class ClassWithNoPublicMethod
+    private readonly string _field;
+
+    public ClassWithNoPublicMethod(string field)
     {
-        private readonly string _field;
-
-        public ClassWithNoPublicMethod(string field)
-        {
-            _field = ToUpper(field);
-        }
-
-        private string ToUpper(string value) => value.ToUpper();
+        _field = ToUpper(field);
     }
+
+    private string ToUpper(string value) => value.ToUpper();
 }
