@@ -1,14 +1,13 @@
-﻿namespace CodeAnalyzer.Tests.Analysis.Cohesion.TestClasses
+﻿namespace CodeAnalyzer.Tests.Analysis.Cohesion.TestClasses;
+
+public class ClassWithOnePublicMethod
 {
-    public class ClassWithOnePublicMethod
+    private readonly string _field;
+
+    public ClassWithOnePublicMethod(string field)
     {
-        private readonly string _field;
-
-        public ClassWithOnePublicMethod(string field)
-        {
-            _field = field;
-        }
-
-        public string ToUpper() => _field.ToUpper();
+        _field = field;
     }
+
+    public string ToUpper() => _field.ToUpper();
 }
