@@ -1,18 +1,17 @@
-﻿namespace CodeAnalyzer.Tests.Analysis.Cohesion.TestClasses
+﻿namespace CodeAnalyzer.Tests.Analysis.Cohesion.TestClasses;
+
+public class ClassWithTwoMethodsAccessingTwoFields
 {
-    public class ClassWithTwoMethodsAccessingTwoFields
+    private readonly string _firstAccessedField;
+    private readonly string _secondAccessedField;
+
+    public ClassWithTwoMethodsAccessingTwoFields(string firstAccessedField, string secondAccessedField)
     {
-        private readonly string _firstAccessedField;
-        private readonly string _secondAccessedField;
-
-        public ClassWithTwoMethodsAccessingTwoFields(string firstAccessedField, string secondAccessedField)
-        {
-            _firstAccessedField = firstAccessedField;
-            _secondAccessedField = secondAccessedField;
-        }
-
-        public string FirstAccess() => _firstAccessedField.ToUpper();
-
-        public string SecondAccess() => _secondAccessedField.ToUpper();
+        _firstAccessedField = firstAccessedField;
+        _secondAccessedField = secondAccessedField;
     }
+
+    public string FirstAccess() => _firstAccessedField.ToUpper();
+
+    public string SecondAccess() => _secondAccessedField.ToUpper();
 }
