@@ -1,16 +1,15 @@
-﻿namespace CodeAnalyzer.Tests.Analysis.Cohesion.TestClasses
+﻿namespace CodeAnalyzer.Tests.Analysis.Cohesion.TestClasses;
+
+public class ClassWithTwoMethodsIncludingADanglingOne
 {
-    public class ClassWithTwoMethodsIncludingADanglingOne
+    private readonly string _firstAccessedField;
+
+    public ClassWithTwoMethodsIncludingADanglingOne(string firstAccessedField)
     {
-        private readonly string _firstAccessedField;
-
-        public ClassWithTwoMethodsIncludingADanglingOne(string firstAccessedField)
-        {
-            _firstAccessedField = firstAccessedField;
-        }
-
-        public string FirstAccess() => _firstAccessedField.ToUpper();
-
-        public string DanglingMethod() => "Some value";
+        _firstAccessedField = firstAccessedField;
     }
+
+    public string FirstAccess() => _firstAccessedField.ToUpper();
+
+    public string DanglingMethod() => "Some value";
 }
