@@ -1,10 +1,9 @@
-﻿namespace CodeAnalyzer.Tests.Analysis.Coupling.TestClasses
+﻿namespace CodeAnalyzer.Tests.Analysis.Coupling.TestClasses;
+
+public class ClassWithDependencyThroughInheritance : SuperClass, ISuperInterface
 {
-    public class ClassWithDependencyThroughInheritance : SuperClass, ISuperInterface
-    {
 
-    }
-
-    public class SuperClass {}
-    public interface ISuperInterface {}
 }
+
+public class SuperClass {}
+public interface ISuperInterface {}
