@@ -1,23 +1,22 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace CodeAnalyzer.Tests.Analysis.Coupling.TestClasses
+namespace CodeAnalyzer.Tests.Analysis.Coupling.TestClasses;
+
+public class ClassWithOnlySystemDependencies
 {
-    public class ClassWithOnlySystemDependencies
+    private readonly int _someInteger;
+
+    public Guid GuidProperty { get; }
+
+    public ClassWithOnlySystemDependencies(int someInteger, Guid guidProperty)
     {
-        private readonly int _someInteger;
+        _someInteger = someInteger;
+        GuidProperty = guidProperty;
+    }
 
-        public Guid GuidProperty { get; }
-
-        public ClassWithOnlySystemDependencies(int someInteger, Guid guidProperty)
-        {
-            _someInteger = someInteger;
-            GuidProperty = guidProperty;
-        }
-
-        public IEnumerable<string> AMethodWithParameters(string stringFromSystem, List<string> list)
-        {
-            return new List<string> {stringFromSystem};
-        }
+    public IEnumerable<string> AMethodWithParameters(string stringFromSystem, List<string> list)
+    {
+        return new List<string> {stringFromSystem};
     }
 }
