@@ -1,82 +1,81 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 
-namespace CodeAnalyzer.Analysis.Coupling
+namespace CodeAnalyzer.Analysis.Coupling;
+
+public class DependencyGraph
 {
-    public class DependencyGraph
+    private readonly Dictionary<string, DependencyNode> _nodesByClassName;
+
+    public int Count => _nodesByClassName.Count;
+    public IEnumerable<DependencyNode> Nodes => _nodesByClassName.Values;
+
+    private DependencyGraph(Dictionary<string,DependencyNode> nodesByClassName)
     {
-        private readonly Dictionary<string, DependencyNode> _nodesByClassName;
-
-        public int Count => _nodesByClassName.Count;
-        public IEnumerable<DependencyNode> Nodes => _nodesByClassName.Values;
-
-        private DependencyGraph(Dictionary<string,DependencyNode> nodesByClassName)
-        {
-            _nodesByClassName = nodesByClassName;
-        }
-
-        public IEnumerable<string> FindDependencies(string className) =>
-            _nodesByClassName.TryGetValue(className, out var node) ? node.Dependencies : Enumerable.Empty<string>();
-
-        public IEnumerable<string> FindReferences(string className) =>
-            _nodesByClassName.TryGetValue(className, out var node) ? node.References : Enumerable.Empty<string>();
-
-        public DependencyNode FindNode(string className) => _nodesByClassName[className];
-
-        public static DependencyGraph FromClasses(IReadOnlyCollection<ClassAnalyzer> classes)
-        {
-            Dictionary<string, IList<string>> dependenciesByClassName = new Dictionary<string, IList<string>>();
-            Dictionary<string, IList<string>> referencesByClassName = new Dictionary<string, IList<string>>();
-
-            Dictionary<string, DependencyNode> nodesByClassName = new Dictionary<string, DependencyNode>();
-
-            foreach (var @class in classes)
-            {
-                var dependencies = @class.FindDependencies().ToList();
-                dependenciesByClassName[@class.QualifiedName] = dependencies;
-
-                foreach (var dependency in dependencies)
-                {
-                    if (!referencesByClassName.ContainsKey(dependency))
-                    {
-                        referencesByClassName[dependency] = new List<string>();
-                    }
-
-                    referencesByClassName[dependency].Add(@class.QualifiedName);
-                }
-            }
-
-            foreach (var @class in classes)
-            {
-                if (!dependenciesByClassName.TryGetValue(@class.QualifiedName, out var dependencies))
-                {
-                    dependencies = new List<string>();
-                }
-
-                if (!referencesByClassName.TryGetValue(@class.QualifiedName, out var references))
-                {
-                    references = new List<string>();
-                }
-
-                nodesByClassName[@class.QualifiedName] = new DependencyNode(@class, dependencies.ToHashSet(), references.ToHashSet());
-            }
-
-            return new DependencyGraph(nodesByClassName);
-        }
+        _nodesByClassName = nodesByClassName;
     }
 
-    public class DependencyNode
-    {
-        public string Identifier => Class.QualifiedName;
-        public IEnumerable<string> Dependencies { get; }
-        public IEnumerable<string> References { get; }
-        public ClassAnalyzer Class { get; }
+    public IEnumerable<string> FindDependencies(string className) =>
+        _nodesByClassName.TryGetValue(className, out var node) ? node.Dependencies : Enumerable.Empty<string>();
 
-        public DependencyNode(ClassAnalyzer @class, IEnumerable<string> dependencies, IEnumerable<string> references)
+    public IEnumerable<string> FindReferences(string className) =>
+        _nodesByClassName.TryGetValue(className, out var node) ? node.References : Enumerable.Empty<string>();
+
+    public DependencyNode FindNode(string className) => _nodesByClassName[className];
+
+    public static DependencyGraph FromClasses(IReadOnlyCollection<ClassAnalyzer> classes)
+    {
+        Dictionary<string, IList<string>> dependenciesByClassName = new();
+        Dictionary<string, IList<string>> referencesByClassName = new();
+
+        Dictionary<string, DependencyNode> nodesByClassName = new();
+
+        foreach (var @class in classes)
         {
-            Dependencies = dependencies;
-            References = references;
-            Class = @class;
+            var dependencies = @class.FindDependencies().ToList();
+            dependenciesByClassName[@class.QualifiedName] = dependencies;
+
+            foreach (var dependency in dependencies)
+            {
+                if (!referencesByClassName.ContainsKey(dependency))
+                {
+                    referencesByClassName[dependency] = new List<string>();
+                }
+
+                referencesByClassName[dependency].Add(@class.QualifiedName);
+            }
         }
+
+        foreach (var @class in classes)
+        {
+            if (!dependenciesByClassName.TryGetValue(@class.QualifiedName, out var dependencies))
+            {
+                dependencies = new List<string>();
+            }
+
+            if (!referencesByClassName.TryGetValue(@class.QualifiedName, out var references))
+            {
+                references = new List<string>();
+            }
+
+            nodesByClassName[@class.QualifiedName] = new DependencyNode(@class, dependencies.ToHashSet(), references.ToHashSet());
+        }
+
+        return new DependencyGraph(nodesByClassName);
+    }
+}
+
+public class DependencyNode
+{
+    public string Identifier => Class.QualifiedName;
+    public IEnumerable<string> Dependencies { get; }
+    public IEnumerable<string> References { get; }
+    public ClassAnalyzer Class { get; }
+
+    public DependencyNode(ClassAnalyzer @class, IEnumerable<string> dependencies, IEnumerable<string> references)
+    {
+        Dependencies = dependencies;
+        References = references;
+        Class = @class;
     }
 }
