@@ -2,53 +2,52 @@
 using CodeAnalyzer.Analysis.Coupling;
 using Microsoft.CodeAnalysis;
 
-namespace CodeAnalyzer.Report
+namespace CodeAnalyzer.Report;
+
+public class AnalysisConfiguration
 {
-    public class AnalysisConfiguration
+    public static readonly AnalysisConfiguration Default = New().Build();
+
+    public Func<DependencyNode, bool> SelectNode { get; }
+    public Func<IFieldSymbol, bool> SelectField { get; }
+    public Func<IPropertySymbol, bool> SelectProperty { get; }
+
+    private AnalysisConfiguration(Func<DependencyNode, bool> selectNode, Func<IFieldSymbol, bool> selectField, Func<IPropertySymbol, bool> selectProperty)
     {
-        public static readonly AnalysisConfiguration Default = New().Build();
+        SelectNode = selectNode;
+        SelectField = selectField;
+        SelectProperty = selectProperty;
+    }
 
-        public Func<DependencyNode, bool> SelectNode { get; }
-        public Func<IFieldSymbol, bool> SelectField { get; }
-        public Func<IPropertySymbol, bool> SelectProperty { get; }
+    public static Builder New() => new();
 
-        private AnalysisConfiguration(Func<DependencyNode, bool> selectNode, Func<IFieldSymbol, bool> selectField, Func<IPropertySymbol, bool> selectProperty)
+    public class Builder
+    {
+        private Func<DependencyNode, bool> _selectNode = _ => true;
+        private Func<IFieldSymbol, bool> _selectField = _ => true;
+        private Func<IPropertySymbol, bool> _selectProperty = _ => true;
+
+        public Builder WhereNodes(Func<DependencyNode, bool> selectNode)
         {
-            SelectNode = selectNode;
-            SelectField = selectField;
-            SelectProperty = selectProperty;
+            _selectNode = selectNode;
+
+            return this;
         }
 
-        public static Builder New() => new Builder();
-
-        public class Builder
+        public Builder WhereFields(Func<IFieldSymbol, bool> selectField)
         {
-            private Func<DependencyNode, bool> _selectNode = node => true;
-            private Func<IFieldSymbol, bool> _selectField = symbol => true;
-            private Func<IPropertySymbol, bool> _selectProperty = symbol => true;
+            _selectField = selectField;
 
-            public Builder WhereNodes(Func<DependencyNode, bool> selectNode)
-            {
-                _selectNode = selectNode;
-
-                return this;
-            }
-
-            public Builder WhereFields(Func<IFieldSymbol, bool> selectField)
-            {
-                _selectField = selectField;
-
-                return this;
-            }
-
-            public Builder WhereProperties(Func<IPropertySymbol, bool> selectProperty)
-            {
-                _selectProperty = selectProperty;
-
-                return this;
-            }
-
-            public AnalysisConfiguration Build() => new AnalysisConfiguration(_selectNode, _selectField, _selectProperty);
+            return this;
         }
+
+        public Builder WhereProperties(Func<IPropertySymbol, bool> selectProperty)
+        {
+            _selectProperty = selectProperty;
+
+            return this;
+        }
+
+        public AnalysisConfiguration Build() => new(_selectNode, _selectField, _selectProperty);
     }
 }
