@@ -1,34 +1,32 @@
-﻿using System.Collections.Generic;
-using System.Globalization;
+﻿using System.Globalization;
 using System.IO;
 using System.Linq;
 using CsvHelper;
 using CsvHelper.Configuration;
 
-namespace CodeAnalyzer.Report.Export
+namespace CodeAnalyzer.Report.Export;
+
+public class AnalysisReportCsvFileWriter
 {
-    public class AnalysisReportCsvFileWriter
+    private readonly AnalysisReport _report;
+
+    public AnalysisReportCsvFileWriter(AnalysisReport report)
     {
-        private readonly AnalysisReport _report;
+        _report = report;
+    }
 
-        public AnalysisReportCsvFileWriter(AnalysisReport report)
-        {
-            _report = report;
-        }
+    public Stream OpenStream()
+    {
+        var csvRecords = _report.ClassesReports.Select(p => p.Value);
 
-        public Stream OpenStream()
-        {
-            IEnumerable<ClassAnalysisReport> csvRecords = _report.ClassesReports.Select(p => p.Value);
+        var memoryStream = new MemoryStream();
+        TextWriter textWriter = new StreamWriter(memoryStream);
+        var csv = new CsvWriter(textWriter, new CsvConfiguration(CultureInfo.InvariantCulture) {HasHeaderRecord = true});
+        csv.WriteRecords(csvRecords);
 
-            var memoryStream = new MemoryStream();
-            TextWriter textWriter = new StreamWriter(memoryStream);
-            var csv = new CsvWriter(textWriter, new CsvConfiguration(CultureInfo.InvariantCulture) {HasHeaderRecord = true});
-            csv.WriteRecords(csvRecords);
+        textWriter.Flush();
+        memoryStream.Position = 0;
 
-            textWriter.Flush();
-            memoryStream.Position = 0;
-
-            return memoryStream;
-        }
+        return memoryStream;
     }
 }
